# ⚡ perf: Hand artisan refactors of new tail call recursion evm