--- a/scripts/perf-fast.sh
+++ b/scripts/perf-fast.sh
@@ -2,6 +2,7 @@
 
 zig build build-evm-runner -Doptimize=ReleaseFast \
   && zig build build-evm-runner-small -Doptimize=ReleaseSmall \
+  && zig build build-evm-runner-call2 -Doptimize=ReleaseFast \
   && zig build build-orchestrator -Doptimize=ReleaseFast \
   && zig-out/bin/orchestrator \
     --compare \