--- a/scripts/perf-slow.sh
+++ b/scripts/perf-slow.sh
@@ -3,6 +3,7 @@
 
 zig build build-evm-runner -Doptimize=ReleaseFast \
   && zig build build-evm-runner-small -Doptimize=ReleaseSmall \
+  && zig build build-evm-runner-call2 -Doptimize=ReleaseFast \
   && zig build build-orchestrator -Doptimize=ReleaseFast \
   && ./zig-out/bin/orchestrator \
     --compare \