--- a/src/evm/evm.zig
+++ b/src/evm/evm.zig
@@ -614,6 +614,7 @@ pub fn is_currently_executing(self: *const Evm) bool {
 pub usingnamespace @import("evm/set_context.zig");
 
 pub usingnamespace @import("evm/call.zig"); // This provides the actual call() implementation
+pub usingnamespace @import("evm/call2.zig"); // This provides the call2() implementation using interpret2
 pub usingnamespace @import("evm/call_contract.zig");
 pub usingnamespace @import("evm/execute_precompile_call.zig");
 pub usingnamespace @import("evm/staticcall_contract.zig");