--- a/src/evm/evm/call.zig
+++ b/src/evm/evm/call.zig
@@ -52,9 +52,9 @@ pub fn call(self: *Evm, params: CallParams) ExecutionError.Error!CallResult {
     // Create host interface from self
     const host = Host.init(self);
 
-    // Determine if this is a top-level call using frame depth
-    // Frame depth of 0 means top-level, regardless of is_executing flag
-    const is_top_level_call = self.current_frame_depth == 0;
+    // Determine if this is a top-level call
+    // A call is top-level only if we're not already executing
+    const is_top_level_call = !self.is_currently_executing();
     Log.debug("[call] is_top_level_call={}, is_executing={}, current_frame_depth={}", .{ is_top_level_call, self.is_currently_executing(), self.current_frame_depth });
     Log.debug("[call] current_output.len at start={}", .{self.current_output.len});
     // Create snapshot for nested calls early to ensure proper revert on any error