## Summary
Optimize call2 interpreter performance by replacing O(n) bytecode scanning with O(1) precomputed mappings.

## Problem
The call2 interpreter was performing O(n) bytecode scanning on every PUSH, JUMP, JUMPI, and PC operation, creating quadratic complexity for contracts with many instructions.

## Solution
Created `SimpleAnalysis` module with precomputed mappings:
- `inst_to_pc`: Maps instruction index → PC value (O(1) lookup)
- `pc_to_inst`: Maps PC → instruction index (O(1) lookup)  
- `push_values`: HashMap of PC → push value (O(1) lookup)

## Performance Impact
**ten-thousand-hashes benchmark:**
- Before: ~30ms
- After: ~4ms
- **~7x faster** 🚀

**erc20-transfer benchmark:**
- Before: ~45ms
- After: ~7ms
- **~6x faster** 🚀

## Implementation Details
1. ✨ Added `SimpleAnalysis` module for precomputed mappings
2. ♻️ Updated Frame struct to store analysis pointer
3. ♻️ Integrated SimpleAnalysis into interpret2
4. ⚡ Optimized tailcall opcodes with cached lookups
5. ♻️ Updated type definitions for consistency

All functions include fallback to O(n) method if analysis is unavailable, ensuring backwards compatibility.

## Testing
- ✅ All existing tests pass
- ✅ Benchmarks show significant performance improvement
- ✅ No functional regressions

🤖 Generated with [Claude Code](https://claude.ai/code)